feat : ServiceLocater, MVP, DIContainer, Installer, Player스탯, Player 허기, 갈증, 체력, 경험치 UI 연동

+++ Assets/02. Scripts/UI/Data/StatusData.cs
@@ -5,11 +5,11 @@ public class StatusData
 {
     public int Level;
     public int EXP;
-    public int Thirst;
-    public int Hunger;
+    public float Thirst;
+    public float Hunger;
     public float HP;
 
-    public StatusData(int level = 1, int exp = 0, int thirst= 100, int hunger = 100, float hp = 100f)
+    public StatusData(int level = 1, int exp = 0, float thirst = 100, float hunger = 100, float hp = 100f)
     {
         Level = level;
         EXP = exp;
